+++ examples/complete/extra_script.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+mkdir -p /home/test-dir
+ls /home