+++ examples/cpu_credits/variables.tf
@@ -16,6 +16,12 @@ variable "instance_type" {
   default     = "t3.large"
 }
 
+variable "architecture" {
+  type        = string
+  description = "The architecture of the instance to launch"
+  default     = "x86_64"
+}
+
 variable "create_ec2_kms_key" {
   description = "Choose whether to create kms key for ebs encryption"
   type        = bool