+++ examples/windows-instance/variables.tf
@@ -22,20 +22,8 @@ variable "create_ec2_kms_key" {
   default     = true
 }
 
-variable "create_key_pair" {
-  description = "Choose whether to create key pair"
-  type        = bool
-  default     = true
-}
-
 variable "ebs_optimized" {
   description = "If true, the launched EC2 instance will be EBS-optimized"
-  type        = bool
-  default     = true
-}
-
-variable "get_password_data" {
-  description = "If true, wait for password data to become available and retrieve it.  Useful for getting the administrator password for instances running Microsoft Windows."
   type        = bool
   default     = true
 }