+++ scripts/userdata.sh
@@ -1,17 +1,86 @@
 #!/bin/bash
 
-DEBUG=$${DEBUG:=off}
+# Install AWS SSM agent for different OS versions
+os_type=$(uname -m)
+os_release=$(grep -oP '(?<=^ID=).+' /etc/os-release | tr -d '"')
 
-if [ $${DEBUG} == 'on' ]; then
-    set -x
+if [ "$${os_type}" == "x86_64" ]; then
+    if [ "$${os_release}" == "amzn" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo yum update -y
+        sudo yum install -y awslogs
+    elif [ "$${os_release}" == "debian" ] || [ "$${os_release}" == "ubuntu" ]; then
+        mkdir /tmp/ssm && cd /tmp/ssm
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/debian_amd64/amazon-ssm-agent.deb
+        sudo dpkg -i amazon-ssm-agent.deb
+        sudo apt-get update -y
+        sudo apt-get install -y awslogs
+    elif [ "$${os_release}" == "centos" ] && [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "8" ] || [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "9" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo dnf update -y
+        sudo dnf install -y awslogs
+    elif [ "$${os_release}" == "centos" ] && [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "7" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo yum update -y
+        sudo yum install -y awslogs
+    elif [ "$${os_release}" == "centos" ] && [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "6" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/3.0.1479.0/linux_amd64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo yum update -y
+        sudo yum install -y awslogs
+    elif [ "$${os_release}" == "rhel" ] && [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "8" ] || [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "9" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo dnf update -y
+        sudo dnf install -y awslogs
+    elif [ "$${os_release}" == "rhel" ] && [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "7" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo yum update -y
+        sudo yum install -y awslogs
+    elif [ "$${os_release}" == "rocky" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo dnf update -y
+        sudo dnf install -y awslogs
+    elif [ "$${os_release}" == "ol" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_amd64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo yum update -y
+        sudo yum install -y awslogs
+    elif [ "$${os_release}" == "sles" ]; then
+        sudo zypper install -y amazon-ssm-agent
+        sudo systemctl start amazon-ssm-agent
+        sudo zypper update -y
+        sudo zypper install -y awslogs
+    fi
+elif [ "$${os_type}" == "aarch64" ]; then
+    if [ "$${os_release}" == "amzn" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_arm64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo yum update -y
+        sudo yum install -y awslogs
+    elif [ "$${os_release}" == "debian" ] || [ "$${os_release}" == "ubuntu" ]; then
+        mkdir /tmp/ssm && cd /tmp/ssm
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/debian_arm64/amazon-ssm-agent.deb
+        sudo dpkg -i amazon-ssm-agent.deb
+        sudo apt-get update -y
+        sudo apt-get install -y awslogs
+    elif [ "$${os_release}" == "rhel" ] && [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "8" ] || [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "9" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_arm64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo dnf update -y
+        sudo dnf install -y awslogs
+    elif [ "$${os_release}" == "rhel" ] && [ "$(grep -oP '(?<=^VERSION_ID=).+' /etc/os-release | tr -d '"' | awk -F. '{print $1}')" == "7" ]; then
+        wget https://s3.amazonaws.com/ec2-downloads-windows/SSMAgent/latest/linux_arm64/amazon-ssm-agent.rpm
+        sudo rpm --install amazon-ssm-agent.rpm
+        sudo yum update -y
+        sudo yum install -y awslogs
+    fi
 fi
 
-TOKEN=$(curl -X PUT "http://169.254.169.254/latest/api/token" -H "X-aws-ec2-metadata-token-ttl-seconds: 300")
-instance_id=$(curl http://169.254.169.254/latest/meta-data/instance-id -H "X-aws-ec2-metadata-token: $TOKEN" | cut -d "-" -f 2)
-container_instance_id=$(curl http://169.254.169.254/latest/meta-data/instance-id -H "X-aws-ec2-metadata-token: $TOKEN")
-
-# Force the update.
-yum update -y
-
-# Software packages:
-yum install -y awslogs vim jq unzip wget tree git
+# Check if the SSM agent is running
+sudo systemctl status amazon-ssm-agent