+++ tests/supportingResources/main.tf
@@ -1,6 +1,7 @@
 module "ec2_vpc" {
+  #checkov:skip=CKV_TF_1: "Ensure Terraform module sources use a commit hash"
   source                 = "boldlink/vpc/aws"
-  version                = "3.0.3"
+  version                = "3.0.4"
   name                   = var.name
   cidr_block             = var.cidr_block
   enable_dns_support     = var.enable_dns_support